refactor: move method `findElement` to `OrchestraContext`
