+++ maestro-orchestra/src/main/java/maestro/orchestra/Orchestra.kt
@@ -19,17 +19,9 @@
 
 package maestro.orchestra
 
-import maestro.DeviceInfo
-import maestro.ElementFilter
-import maestro.Filters
-import maestro.Filters.asFilter
 import maestro.Maestro
 import maestro.MaestroException
 import maestro.MaestroTimer
-import maestro.TreeNode
-import maestro.UiElement
-import maestro.orchestra.filter.FilterWithDescription
-import maestro.orchestra.filter.TraitFilters
 import maestro.orchestra.yaml.YamlCommandReader
 import java.io.File
 import java.nio.file.Files
@@ -117,16 +109,16 @@ class Orchestra(
             is TapOnElementCommand -> {
                 tapOnElement(command, command.retryIfNoChange ?: true, command.waitUntilVisible ?: true)
             }
-            is TapOnPointCommand -> command.execute(OrchestraContext(maestro))
-            is BackPressCommand -> command.execute(OrchestraContext(maestro))
-            is ScrollCommand -> command.execute(OrchestraContext(maestro))
-            is SwipeCommand -> command.execute(OrchestraContext(maestro))
+            is TapOnPointCommand -> command.execute(OrchestraContext(maestro, optionalLookupTimeoutMs, lookupTimeoutMs))
+            is BackPressCommand -> command.execute(OrchestraContext(maestro, optionalLookupTimeoutMs, lookupTimeoutMs))
+            is ScrollCommand -> command.execute(OrchestraContext(maestro, optionalLookupTimeoutMs, lookupTimeoutMs))
+            is SwipeCommand -> command.execute(OrchestraContext(maestro, optionalLookupTimeoutMs, lookupTimeoutMs))
             is AssertCommand -> assertCommand(command)
-            is InputTextCommand -> command.execute(OrchestraContext(maestro))
-            is LaunchAppCommand -> command.execute(OrchestraContext(maestro))
-            is OpenLinkCommand -> command.execute(OrchestraContext(maestro))
-            is PressKeyCommand -> command.execute(OrchestraContext(maestro))
-            is EraseTextCommand -> command.execute(OrchestraContext(maestro))
+            is InputTextCommand -> command.execute(OrchestraContext(maestro, optionalLookupTimeoutMs, lookupTimeoutMs))
+            is LaunchAppCommand -> command.execute(OrchestraContext(maestro, optionalLookupTimeoutMs, lookupTimeoutMs))
+            is OpenLinkCommand -> command.execute(OrchestraContext(maestro, optionalLookupTimeoutMs, lookupTimeoutMs))
+            is PressKeyCommand -> command.execute(OrchestraContext(maestro, optionalLookupTimeoutMs, lookupTimeoutMs))
+            is EraseTextCommand -> command.execute(OrchestraContext(maestro, optionalLookupTimeoutMs, lookupTimeoutMs))
             is ApplyConfigurationCommand, null -> { /* no-op */ }
         }
     }
@@ -139,7 +131,7 @@ class Orchestra(
     private fun assertNotVisible(selector: ElementSelector, timeoutMs: Long = lookupTimeoutMs) {
         val result = MaestroTimer.withTimeout(timeoutMs) {
             try {
-                findElement(selector, timeoutMs = 2000L)
+                OrchestraContext(maestro, optionalLookupTimeoutMs, lookupTimeoutMs).findElement(selector, timeoutMs = 2000L)
 
                 // If we got to that point, the element is still visible.
                 // Returning null to keep waiting.
@@ -160,7 +152,7 @@ class Orchestra(
     }
 
     private fun assertVisible(selector: ElementSelector) {
-        findElement(selector) // Throws if element is not found
+        OrchestraContext(maestro, optionalLookupTimeoutMs, lookupTimeoutMs).findElement(selector) // Throws if element is not found
     }
 
     private fun tapOnElement(
@@ -169,7 +161,7 @@ class Orchestra(
         waitUntilVisible: Boolean,
     ) {
         try {
-            val element = findElement(command.selector)
+            val element = OrchestraContext(maestro, optionalLookupTimeoutMs, lookupTimeoutMs).findElement(command.selector)
             maestro.tap(
                 element,
                 retryIfNoChange,
@@ -184,121 +176,8 @@ class Orchestra(
         }
     }
 
-    private fun findElement(
-        selector: ElementSelector,
-        timeoutMs: Long? = null
-    ): UiElement {
-        val timeout = timeoutMs
-            ?: if (selector.optional) {
-                optionalLookupTimeoutMs
-            } else {
-                lookupTimeoutMs
-            }
-
-        val (description, filterFunc) = buildFilter(
-            selector,
-            maestro.deviceInfo(),
-            maestro.viewHierarchy().aggregate(),
-        )
-
-        return maestro.findElementWithTimeout(
-            timeoutMs = timeout,
-            filter = filterFunc,
-        ) ?: throw MaestroException.ElementNotFound(
-            "Element not found: $description",
-            maestro.viewHierarchy().root,
-        )
-    }
-
-    private fun buildFilter(
-        selector: ElementSelector,
-        deviceInfo: DeviceInfo,
-        allNodes: List<TreeNode>,
-    ): FilterWithDescription {
-        val filters = mutableListOf<ElementFilter>()
-        val descriptions = mutableListOf<String>()
-
-        selector.textRegex
-            ?.let {
-                descriptions += "Text matching regex: $it"
-                filters += Filters.textMatches(it.toRegex(REGEX_OPTIONS)).asFilter()
-            }
-
-        selector.idRegex
-            ?.let {
-                descriptions += "Id matching regex: $it"
-                filters += Filters.idMatches(it.toRegex(REGEX_OPTIONS)).asFilter()
-            }
-
-        selector.size
-            ?.let {
-                descriptions += "Size: $it"
-                filters += Filters.sizeMatches(
-                    width = it.width,
-                    height = it.height,
-                    tolerance = it.tolerance,
-                ).asFilter()
-            }
-
-        selector.below
-            ?.let {
-                descriptions += "Below: ${it.description()}"
-                filters += Filters.below(buildFilter(it, deviceInfo, allNodes).filterFunc)
-            }
-
-        selector.above
-            ?.let {
-                descriptions += "Above: ${it.description()}"
-                filters += Filters.above(buildFilter(it, deviceInfo, allNodes).filterFunc)
-            }
-
-        selector.leftOf
-            ?.let {
-                descriptions += "Left of: ${it.description()}"
-                filters += Filters.leftOf(buildFilter(it, deviceInfo, allNodes).filterFunc)
-            }
-
-        selector.rightOf
-            ?.let {
-                descriptions += "Right of: ${it.description()}"
-                filters += Filters.rightOf(buildFilter(it, deviceInfo, allNodes).filterFunc)
-            }
-
-        selector.containsChild
-            ?.let {
-                descriptions += "Contains child: ${it.description()}"
-                filters += Filters.containsChild(findElement(it)).asFilter()
-            }
-
-        selector.traits
-            ?.map {
-                TraitFilters.buildFilter(it)
-            }
-            ?.forEach { (description, filter) ->
-                descriptions += description
-                filters += filter
-            }
-
-        val finalFilter = selector.index
-            ?.let {
-                Filters.compose(
-                    listOf(
-                        Filters.intersect(filters),
-                        Filters.index(it),
-                    )
-                )
-            } ?: Filters.intersect(filters)
-
-        return FilterWithDescription(
-            descriptions.joinToString(", "),
-            finalFilter,
-        )
-    }
-
     companion object {
-
         val REGEX_OPTIONS = setOf(RegexOption.IGNORE_CASE, RegexOption.DOT_MATCHES_ALL, RegexOption.MULTILINE)
-
     }
 }
